+++ examples/react-app/src/types/contracts/Counter.ts
@@ -5,9 +5,9 @@
 /* eslint-disable @typescript-eslint/consistent-type-imports */
 
 /*
-  Fuels version: 0.94.8
-  Forc version: 0.64.0
-  Fuel-Core version: 0.36.0
+  Fuels version: 0.96.0
+  Forc version: 0.65.2
+  Fuel-Core version: 0.37.1
 */
 
 import { Contract, Interface } from 'fuels';