+++ packages/walletconnect-connector/src/index.ts
@@ -1,2 +1,3 @@
 export * from './WalletConnectConnector';
 export * from './types';
+export * from './utils';