+++ packages/bako-safe/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @fuel-connectors/bako-safe
 
+## 0.35.0
+
+### Minor Changes
+
+- [#387](https://github.com/FuelLabs/fuel-connectors/pull/387) [`5c3d337`](https://github.com/FuelLabs/fuel-connectors/commit/5c3d3375db00b31792a05010a026ed2b74df21db) Thanks [@arboleya](https://github.com/arboleya)! - Upgrading `fuels` to `0.96.0`
+
+## 0.34.2
+
+### Patch Changes
+
+- [#385](https://github.com/FuelLabs/fuel-connectors/pull/385) [`9c0923b`](https://github.com/FuelLabs/fuel-connectors/commit/9c0923b333c35b42cff153182fb85a606dbe0279) Thanks [@helciofranco](https://github.com/helciofranco)! - Revert: fixed node dynamic import error introduced in https://github.com/FuelLabs/fuel-connectors/pull/382
+
+## 0.34.1
+
+### Patch Changes
+
+- [#382](https://github.com/FuelLabs/fuel-connectors/pull/382) [`b0708e3`](https://github.com/FuelLabs/fuel-connectors/commit/b0708e324774a6d95fb0c464433d874c598ce8bd) Thanks [@arthurgeron](https://github.com/arthurgeron)! - Fixed node dynamic import error
+
+## 0.34.0
+
+### Minor Changes
+
+- [#381](https://github.com/FuelLabs/fuel-connectors/pull/381) [`cdac8d4`](https://github.com/FuelLabs/fuel-connectors/commit/cdac8d441bab8e6c7a4c86d60e02c772bdb2e856) Thanks [@luizstacio](https://github.com/luizstacio)! - feat!: change address return to use toString instead of b256
+
+## 0.33.0
+
+## 0.32.0
+
+### Minor Changes
+
+- [#370](https://github.com/FuelLabs/fuel-connectors/pull/370) [`9a20224`](https://github.com/FuelLabs/fuel-connectors/commit/9a202245ceb8653435e963a98fc3453afd81f52b) Thanks [@petertonysmith94](https://github.com/petertonysmith94)! - Updated Fuel SDK to 0.95.0
+
+## 0.31.6
+
+## 0.31.5
+
 ## 0.31.4
 
 ## 0.31.3